# openapi-server/swagger_server/swagger/swagger.yaml
openapi: 3.0.2
info:
  title: Authentication service
  version: "2.0"
servers:
- url: /
paths:
...
  /api/v1/auth/add-personal-data/{user_id}:
    post:
      tags:
      - ADD_PERSONAL_DATA
      summary: Endpoint for user to add personal data
      description: Additional info about user
      operationId: api_v1_auth_add_personal_data_user_id_post
      parameters:
      - name: user_id
        in: path
        description: User id to add/change/delete personal data
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserData'
            example:
              birth_date: "Thu, 08 Oct 1970 00:00:00 GMT"
              city: Cambridge
              email: matt@damon.com
              first_name: Matt
              last_name: Damon
              phone: 71234567
      responses:
        "201":
          description: User data was added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: user personal was data added successfully
                status: success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
        "403":
          description: Access is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: permission denied
                status: error
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: resource not found
                status: error
      security:
      - jwt_auth:
        - "write:admin,subscriber,member"
        - "read:admin,subscriber,member"
      x-openapi-router-controller: swagger_server.controllers.addpersonaldata_controller
  /api/v1/auth/change-password/{user_id}:
    patch:
      tags:
      - CHANGE_PASSWORD
      summary: Endpoint to change forgotten password
      description: Change user password
      operationId: api_v1_auth_change_password_user_id_patch
      parameters:
      - name: user_id
        in: path
        description: User id to change history
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Passwords'
            example:
              new_password: 678910
              old_password: 12345
      responses:
        "200":
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: password changed successfully
                status: success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
        "403":
          description: Access is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: permission denied
                status: error
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: resource not found
                status: error
      security:
      - jwt_auth:
        - "write:admin,subscriber,member"
        - "read:admin,subscriber,member"
      x-openapi-router-controller: swagger_server.controllers.changepassword_controller
  /api/v1/auth/change-personal-data/{user_id}:
    patch:
      tags:
      - CHANGE_PERSONAL_DATA
      summary: Endpoint for user to change data
      description: Additional info about user
      operationId: api_v1_auth_change_personal_data_user_id_patch
      parameters:
      - name: user_id
        in: path
        description: User id to add/change/delete personal data
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserData'
            example:
              birth_date: "Thu, 08 Oct 1970 00:00:00 GMT"
              city: Cambridge
              email: matt@damon.com
              first_name: Matt
              last_name: Damon
              phone: 71234567
      responses:
        "200":
          description: User data was changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: user personal data was changed successfully
                status: success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
        "403":
          description: Access is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: permission denied
                status: error
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: resource not found
                status: error
      security:
      - jwt_auth:
        - "write:admin,subscriber,member"
        - "read:admin,subscriber,member"
      x-openapi-router-controller: swagger_server.controllers.changepersonaldata_controller
  /api/v1/auth/delete-personal-data/{user_id}:
    delete:
      tags:
      - DELETE_PERSONAL_DATA
      summary: Endpoint to delete user personal data
      description: Additional info about user
      operationId: api_v1_auth_delete_personal_data_user_id_delete
      parameters:
      - name: user_id
        in: path
        description: User id to add/change/delete personal data
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: User data was deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: user personal data was deleted successfully
                status: success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
        "403":
          description: Access is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: permission denied
                status: error
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: resource not found
                status: error
      security:
      - jwt_auth:
        - "write:admin,subscriber,member"
        - "read:admin,subscriber,member"
      x-openapi-router-controller: swagger_server.controllers.deletepersonaldata_controller
  /api/v1/auth/login:
    post:
      tags:
      - LOGIN
      summary: Endpoint for user login
      description: Get JWT tokens after login
      operationId: api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
            example:
              password: 12345
              username: yandex
      responses:
        "200":
          description: A pair of access/refresh tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: JWT tokens were generated successfully
                status: success
                tokens:
                  access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
                  refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        "401":
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: username or password are not correct
                status: error
      x-openapi-router-controller: swagger_server.controllers.login_controller
  /api/v1/auth/login-history/{user_id}:
    get:
      tags:
      - LOGIN_HISTORY
      summary: Endoint to get history of user logouts
      description: info about user login
      operationId: api_v1_auth_login_history_user_id_get
      parameters:
      - name: user_id
        in: path
        description: User id to view login history
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: User login history is available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                history:
                - device:
                    created_at: "Mon, 15 Feb 2021 00:00:00 GMT"
                    ip: 89.100.100.100
                    user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36\
                      \ (KHTML, like Gecko) Chrome/97.0.4692.99 Safari/537.36"
                  login_date: "Sun, 06 Feb 2022 00:00:00 GMT"
                - device:
                    created_at: "Mon, 15 Feb 2021 00:00:00 GMT"
                    ip: 89.100.100.100
                    user_agent: "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36\
                      \ (KHTML, like Gecko) Chrome/97.0.4692.99 Safari/537.36"
                  login_date: "Fri, 04 Feb 2022 00:00:00 GMT"
                message: user login history is available
                status: success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
        "403":
          description: Access is not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: permission denied
                status: error
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: resource not found
                status: error
      security:
      - jwt_auth:
        - "write:admin,subscriber,member"
        - "read:admin,subscriber,member"
      x-openapi-router-controller: swagger_server.controllers.loginhistory_controller
  /api/v1/auth/logout:
    post:
      tags:
      - LOGOUT
      summary: Endpoint to logout user
      description: User logout
      operationId: api_v1_auth_logout_post
      responses:
        "200":
          description: Logout successfull
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: User logout successfull
                status: success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
      security:
      - jwt_auth:
        - "write:admin,subscriber,member"
      x-openapi-router-controller: swagger_server.controllers.logout_controller
  /api/v1/auth/refresh-token:
    post:
      tags:
      - REFRESH_TOKEN
      summary: Endoint to refresh expired tokens
      description: Refresh expired tokens
      operationId: api_v1_auth_refresh_token_post
      responses:
        "200":
          description: New tokens were generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: New tokens were generated
                status: success
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              examples:
                expired:
                  value:
                    message: token expired
                    status: error
                notvalid:
                  value:
                    message: token is not valid
                    status: error
      security:
      - jwt_auth:
        - "write:admin,subscriber,member"
      x-openapi-router-controller: swagger_server.controllers.refreshtoken_controller
  /api/v1/auth/register:
    post:
      tags:
      - REGISTRATION
      summary: Endpoint to register new account
      description: Create new user account
      operationId: api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credentials'
            example:
              password: 12345
              username: yandex
      responses:
        "200":
          description: Successfull registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: New account was registered successfully
                status: success
        "400":
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                message: The username is already in use
                status: error
      x-openapi-router-controller: swagger_server.controllers.registration_controller
...
components:
  schemas:
    Credentials:
      title: Credentials
      required:
      - password
      - username
      type: object
      properties:
        password:
          type: string
          writeOnly: true
        username:
          type: string
          writeOnly: true
    Passwords:
      title: Passwords
      required:
      - new_password
      - old_password
      type: object
      properties:
        new_password:
          type: string
          writeOnly: true
        old_password:
          type: string
          writeOnly: true
    Response:
      title: Response
      type: object
      properties:
        message:
          title: response message
          type: string
        status:
          title: response status
          type: string
      example:
        message: message
        status: status
...
    UserData:
      title: UserData
      type: object
      properties:
        birth_date:
          title: birth date
          type: string
        city:
          title: city
          type: string
        email:
          title: email
          type: string
        first_name:
          title: first name
          type: string
        last_name:
          title: last name
          type: string
        phone:
          title: phone
          type: string
...
  securitySchemes:
    jwt_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_jwt_auth